pc-story-dev_sj_lms/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
altgraph==0.17.2
Appium-Python-Client==1.2.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
decorator==4.4.2
future==0.18.2
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
imutils==0.5.4
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
MouseInfo==0.1.3
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.62
packaging==21.3
pefile==2021.9.3
Pillow==8.4.0
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySimpleGUI==4.57.0
pytest==7.0.1
pytest-ordering==0.6
pytest-repeat==0.9.1
python-dateutil==2.8.2
pytweening==1.0.4
PyUserInput==0.1.10
PyWavelets==1.1.1
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.4
selenium==3.141.0
six==1.16.0
tifffile==2020.9.3
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.6.0
